+++ src/index.ts
@@ -33,7 +33,7 @@ export {
   isSymbol,
 } from './lang/object'
 export { isIdCard, isString } from './lang/string'
-export { default as amount } from './amount/amount'
+export { Amout, amount, default as SafeNumber } from './safe-number/safeNumber'
 export { convertToBase26 } from './number/base'
 export type { Unit } from './number/byte'
 export { autoConvertBytes, convertBytes, convertBytesTo } from './number/byte'